refactor(eupago): consolidate shared create/error logic into base class
