+++ src/PayShop/PayShop.php
@@ -3,7 +3,6 @@
 namespace CodeTech\EuPago\PayShop;
 
 use CodeTech\EuPago\EuPago;
-use Illuminate\Support\Facades\Http;
 
 class PayShop extends EuPago
 {
@@ -27,13 +26,6 @@ class PayShop extends EuPago
     protected $id;
 
     /**
-     * The errors stored during the operations.
-     *
-     * @var array
-     */
-    protected $errors = [];
-
-    /**
      * PayShop constructor.
      *
      * @param float $value
@@ -43,57 +35,6 @@ class PayShop extends EuPago
     {
         $this->value = $value;
         $this->id    = $id;
-    }
-
-    /**
-     * Returns the errors.
-     *
-     * @return array
-     */
-    public function getErrors(): array
-    {
-        return $this->errors;
-    }
-
-    /**
-     * Adds an error to the bag.
-     *
-     * @param $code
-     * @param $message
-     */
-    protected function addError($code, $message)
-    {
-        $this->errors[$code] = html_entity_decode($message);
-    }
-
-    /**
-     * Determines whether errors are logged.
-     *
-     * @return bool
-     */
-    public function hasErrors()
-    {
-        return count($this->errors) > 0;
-    }
-
-    /**
-     * Generates a new PayShop reference.
-     *
-     * @return array
-     * @throws \Illuminate\Http\Client\ConnectionException
-     * @throws \Illuminate\Http\Client\RequestException
-     */
-    public function create(): array
-    {
-        $response = Http::asForm()->post($this->getBaseUri() . self::URI, $this->getParams())->throw();
-
-        $referenceData = $response->json();
-
-        if (!$referenceData['sucesso']) {
-            $this->addError($referenceData['estado'], $referenceData['resposta']);
-        }
-
-        return $this->mappedReferenceKeys($referenceData);
     }
 
     /**
